refactor: 重构项目
- 调整项目结构
- 使用 uv 管理依赖
- 使用 bilibili-api-python 获取视频数据/执行搜索
- 使用 curl_cffi 作为 bilibili-api-python 的请求库
