+++ .github/actions/setup-python/action.yml
@@ -1,0 +1,25 @@
+name: Setup Python
+description: Setup Python
+
+inputs:
+  python-version:
+    description: Python version
+    required: false
+    default: "3.13"
+
+runs:
+  using: "composite"
+  steps:
+    - name: Setup uv
+      uses: astral-sh/setup-uv@v5
+      with:
+        python-version: ${{ inputs.python-version }}
+        enable-cache: true
+
+    - name: Install dependencies
+      shell: bash
+      run: uv sync --all-extras --dev
+
+    - name: Configure venv
+      shell: bash
+      run: echo "$PWD/.venv/bin" >> $GITHUB_PATH
