--- requirements.txt
@@ -1,10 +1,0 @@
-requests>=2.32.3
-selenium>=4.30.0
-beautifulsoup4>=4.13.3
-PyQt6>=6.8.1
-aiohttp>=3.11.18
-mutagen>=1.47.0
-tqdm>=4.67.1
-PyQt6-Fluent-Widgets[full]
-darkdetect~=0.8.0
-loguru~=0.7.3
