+++ src/ui/__init__.py
@@ -1,0 +1,1 @@
+from .main_window import MainWindow as MainWindow
