+++ src/utils/player.py
@@ -1,22 +1,25 @@
-import os
 import random
+from pathlib import Path
+from typing import Protocol
 
-from PyQt6.QtCore import Qt, QUrl
 from loguru import logger
+from PyQt6.QtCore import Qt, QUrl
 from qfluentwidgets import InfoBar, InfoBarPosition
 
-import config
-from config import MAIN_PATH, cfg
-from text_tools import remove_before_last_backslash
-from tipbar_tools import open_info_tip
+from src.config import MUSIC_DIR, cfg
+
+from .text import remove_before_last_backslash
+from .tipbar import open_info_tip
 
 
-def open_player():
+def open_player() -> None:
     """打开播放器"""
-    cfg.PLAYER.show()
-    
+    if cfg.PLAYER is not None:
+        cfg.PLAYER.show()
+
+
 def previousSong():
-    """ 播放上一首 """
+    """播放上一首"""
     if cfg.play_queue_index <= 0:
         InfoBar.info(
             "提示",
@@ -24,7 +27,7 @@ def previousSong():
             orient=Qt.Orientation.Horizontal,
             position=InfoBarPosition.BOTTOM_RIGHT,
             duration=1000,
-            parent=InfoBar.desktopView()
+            parent=InfoBar.desktopView(),
         )
         return
     try:
@@ -38,13 +41,14 @@ def previousSong():
             orient=Qt.Orientation.Horizontal,
             position=InfoBarPosition.BOTTOM_RIGHT,
             duration=1000,
-            parent=InfoBar.desktopView()
+            parent=InfoBar.desktopView(),
         )
-    except Exception as e:
-        print(e)
+    except Exception:
+        logger.exception("播放上一首时发生错误")
+
 
 def nextSong():
-    """ 播放下一首 """
+    """播放下一首"""
     if cfg.play_mode == 1:
         if cfg.play_queue_index >= len(cfg.play_queue) - 1:
             InfoBar.info(
@@ -53,13 +57,13 @@ def nextSong():
                 orient=Qt.Orientation.Horizontal,
                 position=InfoBarPosition.BOTTOM_RIGHT,
                 duration=1000,
-                parent=InfoBar.desktopView()
+                parent=InfoBar.desktopView(),
             )
             return
     elif cfg.play_mode == 0:
         if cfg.play_queue_index >= len(cfg.play_queue) - 1:
             cfg.play_queue_index = -1
-    elif  cfg.play_mode == 3:
+    elif cfg.play_mode == 3:
         getRandomIndex()
         playSongByIndex()
         return
@@ -75,7 +79,7 @@ def nextSong():
             orient=Qt.Orientation.Horizontal,
             position=InfoBarPosition.BOTTOM_RIGHT,
             duration=1000,
-            parent=InfoBar.desktopView()
+            parent=InfoBar.desktopView(),
         )
     except Exception as e:
         logger.error(e)
@@ -83,10 +87,20 @@ def nextSong():
 
 # noinspection PyTypeChecker
 def playSongByIndex():
-    file_path = getMusicLocalStr(
-        cfg.play_queue[cfg.play_queue_index])
+    if not cfg.play_queue:
+        InfoBar.error(
+            "错误",
+            "播放队列为空",
+            duration=2000,
+            parent=cfg.MAIN_WINDOW,
+            position=InfoBarPosition.BOTTOM_RIGHT,
+        )
+        return
 
-    url = QUrl.fromLocalFile(file_path)
+    file_path = getMusicLocalStr(cfg.play_queue[cfg.play_queue_index])
+
+    url = QUrl.fromLocalFile(file_path and str(file_path))
+    assert cfg.PLAYER is not None, "播放器未初始化"
     cfg.PLAYER.player.setSource(url)
     cfg.PLAYER.player.play()
 
@@ -95,49 +109,59 @@ def playSongByIndex():
     logger.info(f"当前播放歌曲队列位置：{cfg.play_queue_index}")
     open_info_tip()
 
-def getMusicLocal(fileName):
+
+class _ItemWithText(Protocol):
+    def text(self) -> str: ...
+
+
+def getMusicLocal(item: _ItemWithText | None) -> Path | None:
     """获取音乐文件位置"""
-    if not fileName:
+    if not item:
         return None
 
-    return summonMusicLocal(fileName.text())
+    return summonMusicLocal(item.text())
 
-def getMusicLocalStr(fileName):
+
+def getMusicLocalStr(file_name: str) -> Path | None:
     """获取音乐文件位置(字符串)"""
-    if not fileName:
+    if not file_name:
         return None
 
-    return summonMusicLocal(fileName)
+    return summonMusicLocal(file_name)
 
-def summonMusicLocal(fileName):
+
+def summonMusicLocal(file_name: str) -> Path | None:
     """生成音乐文件路径"""
-    if not fileName:
+    if not file_name:
         return None
 
-    music_dir = os.path.join(MAIN_PATH, "music")
-    file_path = os.path.join(music_dir, fileName)
+    file_path = MUSIC_DIR / file_name
 
-    if not os.path.exists(file_path):
+    if not file_path.exists():
         InfoBar.error(
-            "错误", f"找不到文件: {fileName}",
-            duration=2000, parent=config.cfg.MAIN_WINDOW, position=InfoBarPosition.BOTTOM_RIGHT
+            "错误",
+            f"找不到文件: {file_name}",
+            duration=2000,
+            parent=cfg.MAIN_WINDOW,
+            position=InfoBarPosition.BOTTOM_RIGHT,
         )
         return None
 
     return file_path
 
-def getRandomIndex():
+
+def getRandomIndex() -> None:
     """当处于随机模式时，获取随机的歌曲"""
     index = random.randint(0, len(cfg.play_queue) - 1)
-    while index == cfg.play_queue_index: index = random.randint(0, len(cfg.play_queue) - 1)
+    while index == cfg.play_queue_index:
+        index = random.randint(0, len(cfg.play_queue) - 1)
     cfg.play_queue_index = index
 
-def sequencePlay():
+
+def sequencePlay() -> None:
     """顺序播放"""
     try:
         cfg.play_mode = 0
         playSongByIndex()
-    except Exception as e:
-        logger.error(e)
-
-
+    except Exception:
+        logger.exception("顺序播放时发生错误")
