+++ src/utils/text.py
@@ -1,18 +1,17 @@
 import re
 import time
+from collections.abc import Iterable
+from pathlib import Path
 
 from loguru import logger
 
 
-def contain_text(words_set, text):
+def contain_text(words_set: Iterable[str], text: str) -> bool:
     """检测是否包含内容"""
-    for word in words_set:
-        if word in text:
-            return True
-    return False
+    return any(word in text for word in words_set)
 
 
-def remove_text_after_char(text, after_char):
+def remove_text_after_char(text: str, after_char: str) -> str:
     """删除字符后的文本"""
     index = text.find(after_char)
     if index != -1:
@@ -20,22 +19,23 @@ def remove_text_after_char(text, after_char):
     return text
 
 
-def fileName_process(filename):
+def fix_filename(filename: str) -> str:
     """文件名处理，防止文件名导致的各种问题"""
-    chars_to_remove = ['>', '<', '\\', '/', '*', '|', '?', '\"', '&', ';']
-    for char in filename:
-        if char in chars_to_remove: filename = filename.replace(char, '_')
+    for char in (">", "<", "\\", "/", "*", "|", "?", '"', "&", ";"):
+        filename = filename.replace(char, "_")
     return filename
 
-def count_cn_char(text):
+
+def count_cn_char(text: str) -> int:
     """计算中文字符数"""
     count = 0
     for ch in text:
-        if '\u4e00'<text<'\u9fa5' in ch:
+        if "\u4e00" < text < "\u9fa5" in ch:
             count += 1
     return count
 
-def format_date_str(date):
+
+def format_date_str(date: str) -> str:
     """将爬取的时间格式转换为统一格式(YYYY-MM-DD)"""
     try:
         res = re.search("([0-9]*)-([0-9]*)-([0-9]*)", date)
@@ -44,14 +44,14 @@ def format_date_str(date):
             res = re.search("([0-9]*)-([0-9]*)", date)
             if res is None:
                 if "昨天" in date:
-                    localtime = time.localtime(time.time()-3600*24)
-                    res=f"{localtime.tm_year}-{localtime.tm_mon}-{localtime.tm_mday}"
+                    localtime = time.localtime(time.time() - 3600 * 24)
+                    res = f"{localtime.tm_year}-{localtime.tm_mon}-{localtime.tm_mday}"
                 elif "前天" in date:
-                    localtime = time.localtime(time.time()-3600*24*2)
-                    res=f"{localtime.tm_year}-{localtime.tm_mon}-{localtime.tm_mday}"
+                    localtime = time.localtime(time.time() - 3600 * 24 * 2)
+                    res = f"{localtime.tm_year}-{localtime.tm_mon}-{localtime.tm_mday}"
                 else:
                     localtime = time.localtime(time.time())
-                    res=f"{localtime.tm_year}-{localtime.tm_mon}-{localtime.tm_mday}"
+                    res = f"{localtime.tm_year}-{localtime.tm_mon}-{localtime.tm_mday}"
             else:
                 res = res.group(1, 2)
                 res = f"{localtime.tm_year}-{int(res[0])}-{int(res[1])}"
@@ -67,5 +67,4 @@ def format_date_str(date):
 
 def remove_before_last_backslash(path: str) -> str:
     """删除路径中的最后一个斜杠之前的部分"""
-    parts = path.rsplit("\\", 1)
-    return parts[1] if len(parts) > 1 else path
+    return Path(path).name
